moon, stars, nice sky, daynight cycle setting

+++ shaders/shader.vert
@@ -7,10 +7,10 @@ layout (location = 2) in vec4 colour;
 uniform mat4 uMVP;
 uniform vec3 uChunkPos;
 uniform vec3 uCameraPos;
-uniform int uSkyDarken;
+uniform float uSkyDarken;
 
 out vec2 texCoords;
-out int skyDarken;
+out float skyDarken;
 out vec4 tint;
 
 out float vertexDist;
